Add timestamp attribute to OrderState

+++ fleet_management_api/database/db_models.py
@@ -1,6 +1,6 @@
 import dataclasses
 
-from sqlalchemy import Column, Integer, String, JSON, Boolean
+from sqlalchemy import Column, Integer, String, JSON, Boolean, BigInteger
 from sqlalchemy.orm import Mapped, DeclarativeBase, mapped_column
 
 
@@ -30,7 +30,7 @@ class CarStateDBModel(Base):
     speed: Mapped[int] = Column(Integer)
     fuel: Mapped[int] = Column(Integer)
     position: Mapped[dict] = mapped_column(JSON)
-    timestamp: Mapped[int] = Column(Integer, unique=True)
+    timestamp: Mapped[int] = Column(BigInteger, unique=True)
 
 
 @dataclasses.dataclass
@@ -75,7 +75,7 @@ class OrderStateDBModel(Base):
     id: Mapped[int] = Column(Integer, primary_key=True, unique=True)
     status: Mapped[str] = Column(String)
     order_id: Mapped[int] = Column(Integer)
-    timestamp: Mapped[int] = Column(Integer, unique=True)
+    timestamp: Mapped[int] = Column(BigInteger, unique=True)
 
     @classmethod
     @property
